set up basic routing
