--- components/StepsView.js
@@ -1,34 +1,0 @@
-import React, { Component } from 'react';
-import {
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
-
-class StepsView extends Component {
-  render() {
-    return (
-      <View style={ styles.container }>
-        <Text style={ styles.text }>
-          Steps go here:
-        </Text>
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  text: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-});
-
-export default StepsView;
